Cleaned up the Context useEffects and moved Toasts to the context
